backend/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
bcrypt
python-dotenv